Se corrigen errores en la capa de datos por conversión de tipos, se ajusta modelo, se ajusta GUI de productos.

+++ Tier.Data/DAccesorio.cs
@@ -66,7 +66,7 @@ namespace Tier.Data
                 cmd.Parameters.Add(new MySql.Data.MySqlClient.MySqlParameter("intAccion", uspAcciones.Insertar));
                 this.CargarParametros(cmd, obj);
 
-                obj.idaccesorio = Convert.ToByte(base.CurrentDatabase.ExecuteScalar(cmd));
+                obj.idaccesorio = Convert.ToInt32(base.CurrentDatabase.ExecuteScalar(cmd));
 
                 return obj.idaccesorio > 0;
             }
