+++ Tier.Data/DInsumo.cs
@@ -73,7 +73,7 @@ namespace Tier.Data
                 cmd.Parameters.Add(new MySql.Data.MySqlClient.MySqlParameter("intAccion", uspAcciones.Insertar));
                 this.CargarParametros(cmd, obj);
 
-                obj.idinsumo = Convert.ToByte(base.CurrentDatabase.ExecuteScalar(cmd));
+                obj.idinsumo = Convert.ToInt32(base.CurrentDatabase.ExecuteScalar(cmd));
 
                 return obj.idinsumo > 0;
             }
